+++ src/driveType/CompChassis.h
@@ -41,7 +41,7 @@ public:
 
 	// Overritten from BaseDrive.hpp
 	void driveStraight(double speed) override;
-	double* getEncoderValues() override; // Pointer will be used as an array
+	int* getEncoderValues() override; // Pointer will be used as an array
 	float getGyroYaw() override;
